initial registered slide vis
